change structure
1. read data from cloud insert value based on created_at(primary key)
and beacon_id
2. select all data from database
3. calculate, update final gw conclusion
4. delete old data and insert new data
new created_at will be the time loop run

+++ isensitgw_rssi_total/isensitgw_rssi_total.py
@@ -5,13 +5,15 @@ if api_folder not in sys.path:
 
 from isensit_sql import *
 from isensit_dynamo import *
+
 device = "Acc"
 gws = ["GW1", "GW2", "GW3", "GW4", "GW5"]
 beacons = ["3262", "13591", "5833", "13576", "12931","13305","13320","13500", "13560","13606" ]
 gw_dict = {}
+rssi_table = "rssi_total_table"
+table_name = "maxrssi_table"
 
 def upload_data(id, created_at):
-    table_name = "maxrssi_table"
     db.connect_to_db()
 
     if created_at is not None:
@@ -28,49 +30,36 @@ def upload_data(id, created_at):
 	    row_count = old_json["row_count"]
 	    for gw in gws:
 		gw_dict[gw] = old_json[gw]
-	    print "gw_dict " , gw_dict
+        print "gw_dict " , gw_dict
 
 
         if returned_items is not None:
-	    for item in returned_items:
-		print "ITEMT ", item
-		currentt = item['created_at']
-       		item_info = gws[4]
-            	rssi_max = -1000
-		for gw in gws:
-		    if gw in item:
-			rssi = item[gw]
-			if rssi > rssi_max:
-			    rssi_max = rssi
-			    print 'rssi max ', rssi_max
-			    item_info = gw
-			    print "item_info", item_info
-		gw_dict[item_info] = gw_dict[item_info] + 1
-#            	if 'rssi' in item and 'gatewayID' in item:
-#                    rssi = item['rssi']
-#                    if rssi > rssi_max:
-#                    	rssi_max = rssi
-#			print 'rssi max ', rssi_max
-#                    	item_info = item["gatewayID"]
-#			print "item_info", item_info
-#	    if item_info == "SMULDERS_GW_001":
-#		GW1 = GW1 +1
-#	    elif item_info == "SMULDERS_GW_002":
-#		GW2 = GW1+1
-#	    elif item_info == "SMULDERS_GW_003":
-#		GW3 = GW3+1
-#	    elif item_info == "SMULDERS_GW_004":
-#		GW4 = GW4+1
-#		print "GW4 after ", GW4
-	        dydb.insert_rssi_total2(id, gw_dict, currentt)
-	        dydb.delete_rssi_item(id, currentt)
-            db.insert_max_rssi(id, currentt, gw_dict)
-	    db.delete_acc_beacon_data(table_name,id,row_count)
+	    db.update_rssi_total(id, returned_items)
+	    items = db.read_all_data(rssi_table)
+#	    print "itemts ", items
+	    if items is not None:
+	        for item in items:
+		    print "ITEMT ", item
+		    currentt = item['created_at']
+       		    item_info = gws[4]
+            	    rssi_max = -1000
+		    for gw in gws:
+		    	if gw in item:
+			    rssi = item[gw]
+			    if rssi > rssi_max:
+			    	rssi_max = rssi
+			    	print 'rssi max ', rssi_max
+			    	item_info = gw
+			    	print "item_info", item_info
+		    gw_dict[item_info] = gw_dict[item_info] + 1
 
-	    db.close_db()
-#            print(item_info)
-#	    if "gatewayID" in item_info and "created_at" in item_info and "rssi" in item_info:
-#                succeed = dydb.insert_rssi_total(id, item_info)
+                if dydb.insert_rssi_total(id, gw_dict, created_at):
+#                    dydb.delete_rssi_item(id, returned_items)
+                    db.insert_max_rssi(id, currentt, gw_dict)
+	            db.delete_acc_beacon_data(table_name,id,row_count)
+	            db.delete_all_data(rssi_table)
+
+    db.close_db()
 
 
 
@@ -95,6 +84,3 @@ while True:
 	    time.sleep(60)
     else:
 	print("not working hour")
-
-
-
